requirements.txt
greenlet==1.1.3.post0
psycopg2==2.9.5
python-dotenv==0.21.0
SQLAlchemy==1.4.42
waitress==2.1.2
